--- domain/user.go
@@ -1,8 +1,0 @@
-package domain
-
-type User struct {
-	ID       int64  `json:"id" binding:"required"`
-	Name     string `json:"name" binding:"required"`
-	Email    string `json:"email" binding:"required"`
-	Password string `json:"password" binding:"required"`
-}
