+++ entity/user.go
@@ -1,0 +1,7 @@
+package entity
+
+type User struct {
+	ID    int64  `json:"id"`
+	Name  string `json:"name"`
+	Email string `json:"email"`
+}
